feat: implement authentication & authorization with JWT

+++ app/Actions/Recipe/DestroyRecipe.php
@@ -4,9 +4,11 @@ namespace App\Actions\Recipe;
 
 use App\Models\Recipe;
 use App\Traits\JsonResponseTrait;
+use Illuminate\Auth\AuthenticationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Lorisleiva\Actions\Concerns\AsAction;
 use Throwable;
 
@@ -16,7 +18,13 @@ class DestroyRecipe
 
     public function handle(int $id): void
     {
+        $user = Auth::user();
         $recipe = Recipe::findOrFail($id);
+
+        if ($user->id !== $recipe->user_id) {
+            throw new AuthenticationException(); // FIXME
+        }
+
         $recipe->deleteOrFail();
     }
 
@@ -24,13 +32,13 @@ class DestroyRecipe
     {
         try {
             $this->handle($request->id);
+
+            return $this->success(status_code: 204);
         } catch (ModelNotFoundException $e) {
             return $this->failed('Recipe not found', 404, $e->getMessage());
         } catch (Throwable $e) {
             logger($e);
             return $this->failed('Failed to delete recipe', errors: $e->getMessage());
         }
-
-        return $this->success(status_code: 204);
     }
 }
